moved additional code to shared project
