+++ hellomap-android/Activities/Map/CartoVisJSONActivity.cs
@@ -97,87 +97,6 @@ namespace CartoMobileSample
 			});
 		}
 
-		#region INTERNAL CLASSES
-
-		class MyCartoVisBuilder : CartoVisBuilder
-		{
-			VectorLayer vectorLayer; // vector layer for popups
-			MapView mapView;
-
-			public MyCartoVisBuilder(MapView mapView, VectorLayer vectorLayer)
-			{
-				this.mapView = mapView;
-				this.vectorLayer = vectorLayer;
-			}
-
-			public override void SetCenter(MapPos mapPos)
-			{
-				mapView.SetFocusPos(mapView.Options.BaseProjection.FromWgs84(mapPos), 1.0f);
-			}
-
-			public override void SetZoom(float zoom)
-			{
-				mapView.SetZoom(zoom, 1.0f);
-			}
-
-			public override void AddLayer(Layer layer, Variant attributes)
-			{
-				// Add the layer to the map view
-				mapView.Layers.Add(layer);
-
-				// Check if the layer has info window. In that case will add a custom UTF grid event listener to the layer.
-				Variant infoWindow = attributes.GetObjectElement("infowindow");
-
-				if (infoWindow.Type == VariantType.VariantTypeObject)
-				{
-					TileLayer tileLayer = (TileLayer)layer;
-					tileLayer.UTFGridEventListener = new MyUTFGridEventListener(vectorLayer, infoWindow); ;
-				}
-			}
-		}
-
-		class MyUTFGridEventListener : UTFGridEventListener
-		{
-			VectorLayer vectorLayer;
-			//Variant infoWindow;
-
-			public MyUTFGridEventListener(VectorLayer vectorLayer, Variant infoWindow)
-			{
-				this.vectorLayer = vectorLayer;
-				//this.infoWindow = infoWindow;
-			}
-
-			public override bool OnUTFGridClicked(UTFGridClickInfo clickInfo)
-			{
-				LocalVectorDataSource vectorDataSource = (LocalVectorDataSource)vectorLayer.DataSource;
-
-				// Clear previous popups
-				vectorDataSource.Clear();
-
-				// Multiple vector elements can be clicked at the same time, we only care about the one
-				// Check the type of vector element
-				BalloonPopup clickPopup = null;
-				BalloonPopupStyleBuilder styleBuilder = new BalloonPopupStyleBuilder();
-
-				// Configure style
-				styleBuilder.LeftMargins = new BalloonPopupMargins(0, 0, 0, 0);
-				styleBuilder.TitleMargins = new BalloonPopupMargins(6, 3, 6, 3);
-
-				// Make sure this label is shown on top all other labels
-				styleBuilder.PlacementPriority = 10;
-
-				// Show clicked element variant as JSON string
-				string desc = clickInfo.ElementInfo.ToString();
-
-				clickPopup = new BalloonPopup(clickInfo.ClickPos, styleBuilder.BuildStyle(), "Clicked", desc);
-
-				vectorDataSource.Add(clickPopup);
-
-				return true;
-			}
-		}
-
-		#endregion
 	}
 }
